任务新增字段和限制

+++ src/main/java/com/bluewhite/production/group/action/GroupAction.java
@@ -1,21 +1,16 @@
 package com.bluewhite.production.group.action;
 
-import static org.hamcrest.CoreMatchers.nullValue;
-
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.apache.poi.poifs.storage.ListManagedBlock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
@@ -27,11 +22,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.support.ByteArrayMultipartFileEditor;
 
-import com.alibaba.druid.sql.visitor.functions.If;
 import com.bluewhite.basedata.entity.BaseData;
 import com.bluewhite.common.BeanCopyUtils;
 import com.bluewhite.common.ClearCascadeJSON;
-import com.bluewhite.common.Constants;
 import com.bluewhite.common.DateTimePattern;
 import com.bluewhite.common.Log;
 import com.bluewhite.common.SessionManager;
@@ -42,14 +35,17 @@ import com.bluewhite.common.utils.DatesUtil;
 import com.bluewhite.common.utils.UnUtil;
 import com.bluewhite.finance.attendance.dao.AttendancePayDao;
 import com.bluewhite.finance.attendance.entity.AttendancePay;
-import com.bluewhite.finance.attendance.service.AttendancePayService;
 import com.bluewhite.personnel.attendance.entity.Attendance;
 import com.bluewhite.personnel.attendance.service.AttendanceService;
+import com.bluewhite.production.farragotask.entity.FarragoTask;
+import com.bluewhite.production.farragotask.service.FarragoTaskService;
 import com.bluewhite.production.group.dao.TemporarilyDao;
 import com.bluewhite.production.group.entity.Group;
 import com.bluewhite.production.group.entity.Temporarily;
 import com.bluewhite.production.group.service.GroupService;
 import com.bluewhite.production.productionutils.constant.ProTypeUtils;
+import com.bluewhite.production.task.entity.Task;
+import com.bluewhite.production.task.service.TaskService;
 import com.bluewhite.system.user.entity.TemporaryUser;
 import com.bluewhite.system.user.entity.User;
 import com.bluewhite.system.user.service.UserService;
@@ -64,20 +60,21 @@ public class GroupAction {
 	@Autowired
 	private UserService userService;
 	@Autowired
+	private TaskService taskService;
+	@Autowired
 	private TemporarilyDao temporarilyDao;
 	@Autowired
 	private AttendancePayDao attendancePayDao;
 	@Autowired
 	private AttendanceService attendanceService;
-	
-	
+	@Autowired
+	private FarragoTaskService farragoTaskService;
 
 	private ClearCascadeJSON clearCascadeJSON;
 	{
 		clearCascadeJSON = ClearCascadeJSON.get()
-				.addRetainTerm(Group.class, "id", "name", "type", "users", "temporaryUsers","kindWork", "remark")
-				.addRetainTerm(User.class, "id", "userName")
-				.addRetainTerm(TemporaryUser.class, "id", "userName")
+				.addRetainTerm(Group.class, "id", "name", "type", "users", "temporaryUsers", "kindWork", "remark")
+				.addRetainTerm(User.class, "id", "userName").addRetainTerm(TemporaryUser.class, "id", "userName")
 				.addRetainTerm(BaseData.class, "id", "name", "type");
 	}
 
@@ -139,7 +136,7 @@ public class GroupAction {
 	 */
 	@RequestMapping(value = "/production/allGroup", method = RequestMethod.GET)
 	@ResponseBody
-	public CommonResponse allGroup(HttpServletRequest request,Long id, Date temporarilyDate) {
+	public CommonResponse allGroup(HttpServletRequest request, Long id, Date temporarilyDate) {
 		CommonResponse cr = new CommonResponse();
 		CurrentUser cu = SessionManager.getUserSession();
 		String sourceMachine = "";
@@ -149,137 +146,141 @@ public class GroupAction {
 		if (cu.getRole().contains("packScene")) {
 			sourceMachine = "ELEVEN_WAREHOUSE";
 		}
-		Map<String,Object> groupMap = new HashMap<>();
+		Map<String, Object> groupMap = new HashMap<>();
 		Date startTime = DatesUtil.getfristDayOftime(ProTypeUtils.countAllotTime(temporarilyDate));
 		Date endTime = DatesUtil.getLastDayOftime(ProTypeUtils.countAllotTime(temporarilyDate));
 		Group group = groupService.findOne(id);
-		List<Temporarily> temporarilyList = temporarilyDao.findByTypeAndGroupIdAndTemporarilyDate(group.getType(),id,startTime);
+		List<Temporarily> temporarilyList = temporarilyDao.findByTypeAndGroupIdAndTemporarilyDate(group.getType(), id,
+				startTime);
 		List<Map<String, Object>> userList = new ArrayList<>();
 		List<Map<String, Object>> temporarilyUserList = new ArrayList<>();
-		//平板模式下，按打卡记录显示正式工作人员
-		if(UnUtil.isFromMobile(request)){
+		// 平板模式下，按打卡记录显示正式工作人员
+		if (UnUtil.isFromMobile(request)) {
 			Date startTimeSix = DatesUtil.dayTime(startTime, "05:00:00");
 			List<User> userGroupList = userService.findByGroupId(id);
 			String sourceMachineFinal = sourceMachine;
-			userGroupList = userGroupList.stream().filter(user->{
-				List<Attendance>  attendanceList = attendanceService.findByUserIdAndSourceMachineAndTimeBetween(user.getId(), sourceMachineFinal, startTimeSix, endTime);
-				if(attendanceList.size()>0){
+			userGroupList = userGroupList.stream().filter(user -> {
+				List<Attendance> attendanceList = attendanceService.findByUserIdAndSourceMachineAndTimeBetween(
+						user.getId(), sourceMachineFinal, startTimeSix, endTime);
+				if (attendanceList.size() > 0) {
 					return true;
-				}else{
+				} else {
 					return false;
 				}
 			}).collect(Collectors.toList());
-			
+
 			if (temporarilyList.size() > 0) {
-				for (Temporarily temporarily : temporarilyList) {	
+				for (Temporarily temporarily : temporarilyList) {
 					// 查询出该分组临时员工的出勤数据
 					if (temporarily.getTemporaryUserId() != null) {
-						Map<String, Object>  temporarilyUserMap = new HashMap<>();
+						Map<String, Object> temporarilyUserMap = new HashMap<>();
 						temporarilyUserMap.put("id", temporarily.getTemporaryUserId());
 						temporarilyUserMap.put("userId", temporarily.getId());
 						temporarilyUserMap.put("secondment", 0);
 						temporarilyUserMap.put("groupId", temporarily.getGroupId());
-						temporarilyUserMap.put("name",temporarily.getTemporaryUser().getUserName());
-						temporarilyUserMap.put("time",temporarily.getWorkTime()==null?0:temporarily.getWorkTime());
-						//工作休息状态
-						temporarilyUserMap.put("status",1);
+						temporarilyUserMap.put("name", temporarily.getTemporaryUser().getUserName());
+						temporarilyUserMap.put("time",
+								temporarily.getWorkTime() == null ? 0 : temporarily.getWorkTime());
+						// 工作休息状态
+						temporarilyUserMap.put("status", 1);
 						temporarilyUserList.add(temporarilyUserMap);
 					}
 					// 查询出该分组本厂借调员工的出勤数据
 					if (temporarily.getUserId() != null) {
-						Map<String, Object>  userMap = new HashMap<>();
+						Map<String, Object> userMap = new HashMap<>();
 						userMap.put("id", temporarily.getUserId());
 						userMap.put("userId", temporarily.getId());
 						userMap.put("secondment", 0);
 						userMap.put("groupId", temporarily.getGroupId());
-						userMap.put("name",temporarily.getUser().getUserName());
-						userMap.put("time",temporarily.getWorkTime()==null?0:temporarily.getWorkTime());
-						userMap.put("status",1);
+						userMap.put("name", temporarily.getUser().getUserName());
+						userMap.put("time", temporarily.getWorkTime() == null ? 0 : temporarily.getWorkTime());
+						userMap.put("status", 1);
 						userList.add(userMap);
 					}
-				} 
+				}
 				groupMap.put("temporarilyUser", temporarilyUserList);
 			}
-			
+
 			if (userGroupList.size() > 0) {
 				// 查询出该分组本厂员工的出勤数据
 				for (User user : userGroupList) {
 					// 获取员工一天内的打卡记录并按照自然排序
-					List<Attendance>  attendanceList = attendanceService.findByUserIdAndSourceMachineAndTimeBetween(user.getId(),sourceMachine, startTime, endTime)
-							.stream().sorted(Comparator.comparing(Attendance::getTime))
-							.collect(Collectors.toList());
+					List<Attendance> attendanceList = attendanceService
+							.findByUserIdAndSourceMachineAndTimeBetween(user.getId(), sourceMachine, startTime, endTime)
+							.stream().sorted(Comparator.comparing(Attendance::getTime)).collect(Collectors.toList());
 					Attendance attendanceIn = null;
 					Attendance attendanceOut = null;
 					Double time = null;
 					Double timeH = null;
 					int flag = 0;
-					//根据签到时间实时显示工作时长
-					if(attendanceList.size()>0){
+					// 根据签到时间实时显示工作时长
+					if (attendanceList.size() > 0) {
 						flag = 1;
 						attendanceIn = attendanceList.get(0);
 						time = DatesUtil.getTime(attendanceIn.getTime(), new Date());
-						timeH =DatesUtil.getTimeHourPick(attendanceIn.getTime(),new Date());
+						timeH = DatesUtil.getTimeHourPick(attendanceIn.getTime(), new Date());
 					}
-					if(attendanceIn.getManualTime()!=null){
+					if (attendanceIn.getManualTime() != null) {
 						flag = 0;
 						time = DatesUtil.getTime(attendanceIn.getTime(), attendanceIn.getManualTime());
-						timeH =DatesUtil.getTimeHourPick(attendanceIn.getTime(), attendanceIn.getManualTime());
+						timeH = DatesUtil.getTimeHourPick(attendanceIn.getTime(), attendanceIn.getManualTime());
 					}
-					Map<String, Object>  userMap = new HashMap<>();
+					Map<String, Object> userMap = new HashMap<>();
 					userMap.put("id", user.getId());
 					userMap.put("userId", attendanceIn.getId());
 					userMap.put("secondment", 1);
 					userMap.put("groupId", user.getGroupId());
-					userMap.put("name",user.getUserName());
-					userMap.put("time",timeH+"H/("+time+"M)");
-					userMap.put("status",flag);
+					userMap.put("name", user.getUserName());
+					userMap.put("time", timeH + "H/(" + time + "M)");
+					userMap.put("status", flag);
 					userList.add(userMap);
 				}
 				groupMap.put("userList", userList);
 			}
-		}else{
-			//pc模式下
-			List<AttendancePay> attendancePayList = attendancePayDao.findByGroupIdAndTypeAndAllotTimeBetween(id,group.getType(), startTime, endTime);
+		} else {
+			// pc模式下
+			List<AttendancePay> attendancePayList = attendancePayDao.findByGroupIdAndTypeAndAllotTimeBetween(id,
+					group.getType(), startTime, endTime);
 			if (temporarilyList.size() > 0) {
 				for (Temporarily temporarily : temporarilyList) {
 					// 查询出该分组临时员工的出勤数据
 					if (temporarily.getTemporaryUserId() != null) {
-						Map<String, Object>  temporarilyUserMap = new HashMap<>();
+						Map<String, Object> temporarilyUserMap = new HashMap<>();
 						temporarilyUserMap.put("id", temporarily.getTemporaryUserId());
 						temporarilyUserMap.put("userId", temporarily.getId());
 						temporarilyUserMap.put("secondment", 0);
 						temporarilyUserMap.put("groupId", temporarily.getGroupId());
-						temporarilyUserMap.put("name",temporarily.getTemporaryUser().getUserName());
-						temporarilyUserMap.put("time",temporarily.getWorkTime());
-						temporarilyUserMap.put("status",1);
+						temporarilyUserMap.put("name", temporarily.getTemporaryUser().getUserName());
+						temporarilyUserMap.put("time", temporarily.getWorkTime());
+						temporarilyUserMap.put("status", 1);
 						temporarilyUserList.add(temporarilyUserMap);
 					}
 					// 查询出该分组本厂借调员工的出勤数据
 					if (temporarily.getUserId() != null) {
-						Map<String, Object>  userMap = new HashMap<>();
+						Map<String, Object> userMap = new HashMap<>();
 						userMap.put("id", temporarily.getUserId());
 						userMap.put("userId", temporarily.getId());
 						userMap.put("secondment", 0);
 						userMap.put("groupId", temporarily.getGroupId());
-						userMap.put("name",temporarily.getUser().getUserName());
-						userMap.put("time",temporarily.getWorkTime());
-						userMap.put("status",1);
+						userMap.put("name", temporarily.getUser().getUserName());
+						userMap.put("time", temporarily.getWorkTime());
+						userMap.put("status", 1);
 						userList.add(userMap);
 					}
-				} 
+				}
 				groupMap.put("temporarilyUser", temporarilyUserList);
 			}
 			if (attendancePayList.size() > 0) {
 				// 查询出该分组本厂员工的出勤数据
 				for (AttendancePay attendancePay : attendancePayList) {
-					Map<String, Object>  userMap = new HashMap<>();
+					Map<String, Object> userMap = new HashMap<>();
 					userMap.put("id", attendancePay.getUserId());
 					userMap.put("userId", attendancePay.getId());
 					userMap.put("secondment", 1);
 					userMap.put("groupId", attendancePay.getGroupId());
-					userMap.put("name",attendancePay.getUserName());
-					userMap.put("time",attendancePay.getWorkTime());
-					userMap.put("status",1);
+					userMap.put("name", attendancePay.getUserName());
+					userMap.put("time", attendancePay.getWorkTime());
+					userMap.put("status", 1);
 					userList.add(userMap);
 				}
 				groupMap.put("userList", userList);
@@ -290,23 +291,23 @@ public class GroupAction {
 		cr.setMessage("查询成功");
 		return cr;
 	}
-	
+
 	/**
 	 * 手动签出
 	 * 
 	 */
 	@RequestMapping(value = "/production/updateManualTime", method = RequestMethod.GET)
 	@ResponseBody
-	public CommonResponse updateManualTime(Long id,Integer status,Date time) {
+	public CommonResponse updateManualTime(Long id, Integer status, Date time) {
 		CommonResponse cr = new CommonResponse();
-		if(id!=null){
+		if (id != null) {
 			Attendance attendance = attendanceService.findOne(id);
-			//休息
-			if(status==0){
+			// 休息
+			if (status == 0) {
 				attendance.setManualTime(time);
 			}
-			//工作
-			if(status==1){
+			// 工作
+			if (status == 1) {
 				attendance.setManualTime(null);
 			}
 			attendanceService.save(attendance);
@@ -314,13 +315,9 @@ public class GroupAction {
 		cr.setMessage("修改成功");
 		return cr;
 	}
-	
-	
-	
 
 	/**
-	 * 查询分组
-	 * type 分组所属部门类型 (1=一楼质检，2=一楼包装，3=二楼针工)
+	 * 查询分组 type 分组所属部门类型 (1=一楼质检，2=一楼包装，3=二楼针工)
 	 * 
 	 */
 	@RequestMapping(value = "/production/getGroup", method = RequestMethod.GET)
@@ -344,7 +341,7 @@ public class GroupAction {
 	 */
 	@RequestMapping(value = "/production/userGroup", method = RequestMethod.POST)
 	@ResponseBody
-	public CommonResponse userGroup(String userIds,Long groupId) {
+	public CommonResponse userGroup(String userIds, Long groupId) {
 		CommonResponse cr = new CommonResponse();
 		if (userIds != null) {
 			String[] userIdString = userIds.split(",");
@@ -414,7 +411,8 @@ public class GroupAction {
 			BeanCopyUtils.copyNotEmpty(temporarily, temporarilyNew, "");
 			temporarilyNew.setTemporarilyDate(DatesUtil.getfristDayOftime(date));
 			if (temporarily.getUserId() != null) {
-				if (temporarilyDao.findByUserIdAndTemporarilyDateAndTypeAndGroupId(temporarily.getUserId(),date, temporarily.getType(),temporarily.getGroupId()) != null) {
+				if (temporarilyDao.findByUserIdAndTemporarilyDateAndTypeAndGroupId(temporarily.getUserId(), date,
+						temporarily.getType(), temporarily.getGroupId()) != null) {
 					cr.setMessage("当天当前分组已添加过正式人员的工作时间,不必再次添加");
 					cr.setCode(ErrorCode.ILLEGAL_ARGUMENT.getCode());
 					return cr;
@@ -438,12 +436,13 @@ public class GroupAction {
 	@ResponseBody
 	public CommonResponse updateTemporarily(Temporarily temporarily) {
 		CommonResponse cr = new CommonResponse();
-		if (temporarily.getId()!=null) {
+		if (temporarily.getId() != null) {
 			Temporarily oldtemporarily = temporarilyDao.findOne(temporarily.getId());
 			BeanCopyUtils.copyNotEmpty(temporarily, oldtemporarily);
 			temporarilyDao.save(oldtemporarily);
 			cr.setMessage("修改成功");
-		}else{
+		} else {
+			cr.setCode(ErrorCode.ILLEGAL_ARGUMENT.getCode());
 			cr.setMessage("不能为空");
 		}
 		return cr;
@@ -459,14 +458,12 @@ public class GroupAction {
 	@ResponseBody
 	public CommonResponse getTemporarily(Integer type, Date temporarilyDate) {
 		CommonResponse cr = new CommonResponse();
-		List<Temporarily> temporarilyList = temporarilyDao.findByTypeAndTemporarilyDate(type, temporarilyDate)
-				.stream().filter(Temporarily->Temporarily.getUserId()!=null).collect(Collectors.toList());
+		List<Temporarily> temporarilyList = temporarilyDao.findByTypeAndTemporarilyDate(type, temporarilyDate).stream()
+				.filter(Temporarily -> Temporarily.getUserId() != null).collect(Collectors.toList());
 		cr.setData(ClearCascadeJSON.get()
-				.addRetainTerm(Temporarily.class, "id","workTime", "temporarilyDate", "groupName", "group","user")
-				.addRetainTerm(User.class, "id","userName")
-				.addRetainTerm(Group.class,"id","name")
-				.format(temporarilyList)
-				.toJSON());
+				.addRetainTerm(Temporarily.class, "id", "workTime", "temporarilyDate", "groupName", "group", "user")
+				.addRetainTerm(User.class, "id", "userName").addRetainTerm(Group.class, "id", "name")
+				.format(temporarilyList).toJSON());
 		cr.setMessage("查询成功");
 		return cr;
 	}
@@ -484,9 +481,7 @@ public class GroupAction {
 		List<Temporarily> temporarilyList = groupService.findTemporarilyList(temporarily);
 		cr.setData(ClearCascadeJSON.get()
 				.addRetainTerm(Temporarily.class, "id", "workTime", "workTimeSlice", "temporarilyDate", "group", "user")
-				.addRetainTerm(User.class, "userName")
-				.addRetainTerm(Group.class, "name")
-				.format(temporarilyList)
+				.addRetainTerm(User.class, "userName").addRetainTerm(Group.class, "name").format(temporarilyList)
 				.toJSON());
 		cr.setMessage("查询成功");
 		return cr;
@@ -519,12 +514,26 @@ public class GroupAction {
 		int count = 0;
 		if (!StringUtils.isEmpty(ids)) {
 			String[] idString = ids.split(",");
-			for(String id : idString){
-				temporarilyDao.delete(Long.parseLong(id));
-				count++;
+			for (String id : idString) {
+				Temporarily temporarily = temporarilyDao.findOne(Long.parseLong(id));
+				if (temporarily != null) {
+					Date orderTimeBegin = DatesUtil.getfristDayOftime(temporarily.getTemporarilyDate());
+					Date orderTimeEnd = DatesUtil.getLastDayOftime(temporarily.getTemporarilyDate());
+					List<Task> taskList = taskService.findInSetTemporaryIds(id, orderTimeBegin, orderTimeEnd);
+					List<FarragoTask> farragoTaskList = farragoTaskService.findInSetIds(id, orderTimeBegin,
+							orderTimeEnd);
+					if (taskList.size() > 0 || farragoTaskList.size() > 0) {
+						cr.setCode(ErrorCode.ILLEGAL_ARGUMENT.getCode());
+						cr.setMessage(temporarily.getUser().getUserName() + "当天考勤已分配任务，无法删除，需删除，请先删除任务");
+					}else{
+						temporarilyDao.delete(Long.parseLong(id));
+						count++;
+					}
+				}
+
 			}
 		}
-		cr.setMessage("成功删除"+count+"个");
+		cr.setMessage("成功删除" + count + "条出勤记录");
 		return cr;
 	}
